backend/requirements.txt
asgiref==3.10.0
black==25.1.0
certifi==2025.10.5
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.3.0
colorama==0.4.6
distlib==0.4.0
Django==5.2.7
django-cors-headers==4.9.0
django-stubs==5.1.3
django-stubs-ext==5.2.7
djangorestframework==3.16.1
djangorestframework-stubs==3.15.3
drf-nested-routers==0.94.2
filelock==3.18.0
flake8==7.2.0
identify==2.6.15
idna==3.11
isort==6.0.1
mccabe==0.7.0
mypy==1.18.2
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.7
pycodestyle==2.13.0
pyflakes==3.3.2
PyYAML==6.0.2
requests==2.32.5
sqlparse==0.5.3
types-PyYAML==6.0.12.20250402
types-requests==2.32.4.20250913
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.30.0
pandas==2.3.3
google-generativeai==0.5.4
numpy==1.26.4
google_genai==1.12.1
openai==1.82.0
pydantic==2.11.10
python-dotenv==1.1.1
pytest==8.4.2
pytest-django==4.9.0
pytest-mock==3.14.0
pytest-cov==6.0.0
httpx==0.28.1
opentelemetry-api==1.39.0
opentelemetry-sdk==1.39.0
logfire==4.16.0
PyPDF2==3.0.1
python-docx==1.1.0
apsw>=3.51.1.0
sqlite-vec>=0.1.6